[feature][cvte] 完善移动端界面适配与退出交互
[原因分析] 移动端缺少横屏安全区适配及退出确认，存在弹窗叠加、空字体和技能参数类型问题
[影响范围]
1. 技术栈：安卓
2. 影响模块：移动端横屏与安全区适配、弹窗管理、退出交互、图鉴列表刷新、数字字体、技能战斗、安卓构建配置
3. 是否平台问题：否
4. 风险程度：中

[bug_id] none

+++ Assets/Scripts/技能系统/技能战斗组件.cs
@@ -334,7 +334,7 @@ public class 技能战斗组件 : MonoBehaviour
 		else if (name == "神·徐庶")
 		{
 			清除我方负面();
-			使我方添加状态("清障", value, 0.0);
+			使我方添加状态("清障", (float)value, 0.0);
 		}
 		else if (name == "神·二乔")
 		{
